@definitelytyped/header-parser: move @types/parsimmon dep to devDependencies
Improves compatibility for dependents that have differing global declarations for `require`.

The dependency on `@types/node` causes that package to be installed into `node_modules/` of any dependent.
This is problematic for type definitions that aim for compatibility with TypeScript 2.8, as `@types/parsimmon` triggers TS error TS2370 (A rest parameter must be of an array type).